First RBM Business Directory
Had Claude whip up this little prototype. "We" could make a website that has a huge searchable/sortable/filterable table with all the businesses that use RCS Business Messaging.
Or "we" can just upload an Excel workbook to this repo and "all" commit to it.
Don't know, I've been busy lately.